pyproject.toml
[project]
name = "csat-api"
version = "0.1.0"
description = "CSAT Feedback API"
authors = [
    {name = "Blesson Joel",email = "blessonjoel2004@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "boto3 (>=1.42.49,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "cryptography (>=46.0.5,<47.0.0)",
    "bcrypt (==3.2.2)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
